refactor: refactor code style, fix problems
