+++ package.json
@@ -13,6 +13,7 @@
     "node-sass": "8.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-normalize": "^0.4.4",
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.8.1",
     "react-slick": "^0.29.0",
