+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import Home from './components/Home/Home';
+import { Home } from './components/Home/Home';
 
 const App = () => {
  return (
