+++ src/UI/Input/Input.module.scss
@@ -1,0 +1,17 @@
+.input {
+ background: url('../../assets/loop.svg'), none;
+ background-repeat: no-repeat;
+ background-position: 98% 10px;
+ width: 525px;
+ height: 36px;
+ border: 2px solid #d1d5db;
+ border-radius: 6px;
+ padding-left: 6px;
+ font-weight: 400;
+ font-size: 16px;
+ line-height: 24px;
+ color: #ffffff;
+ &::placeholder {
+  color: #ffffff;
+ }
+}
