+++ src/UI/Input/Input.tsx
@@ -1,3 +1,9 @@
-export const Input = ({ placeholder }: any) => {
- return <input placeholder={placeholder} className="input input--search" />;
+import s from './Input.module.scss';
+
+type Input = {
+ placeholder: string;
+};
+
+export const Input = ({ placeholder }: Input) => {
+ return <input placeholder={placeholder} className={s.input} />;
 };
