+++ src/components/Container/Container.module.scss
@@ -1,0 +1,5 @@
+.container {
+ max-width: 1240px;
+ padding: 0 15px;
+ margin: 0 auto;
+}
