+++ src/components/Header/Header.module.scss
@@ -1,0 +1,8 @@
+.header {
+ padding: 15px 0;
+ &_wrapper {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+ }
+}
