+++ src/components/Header/Header.tsx
@@ -1,13 +1,14 @@
-//container
 import { Container } from '../index';
-//Header
-import { HeaderHome, HeaderForm, HeaderController } from '../index';
+import { HeaderHome } from './HeaderHome/HeaderHome';
+import { HeaderForm } from './HeaderForm/HeaderForm';
+import { HeaderController } from './HeaderController/HeaderController';
+import s from './Header.module.scss';
 
 export const Header = () => {
  return (
-  <div className="header">
+  <div className={s.header}>
    <Container>
-    <div className="header-wrapper">
+    <div className={s.header_wrapper}>
      <HeaderHome />
      <HeaderForm />
      <HeaderController />
