+++ src/components/Header/HeaderControl/HeaderControl.module.scss
@@ -1,0 +1,11 @@
+.header_control {
+ font-weight: 700;
+ font-size: 16px;
+ line-height: 24px;
+ color: #ffffff;
+ cursor: pointer;
+ transition: 0.3s;
+ &:hover {
+  color: #be123c;
+ }
+}
