+++ src/components/Header/HeaderControl/HeaderControl.tsx
@@ -1,3 +1,10 @@
-export const HeaderControl = ({ text, action }: any) => {
- return <span className="header-control">{text}</span>;
+import s from './HeaderControl.module.scss';
+
+type HeaderControl = {
+ text: string;
+ action: string;
+};
+
+export const HeaderControl = ({ text, action }: HeaderControl) => {
+ return <span className={s.header_control}>{text}</span>;
 };
