+++ src/components/Header/HeaderController/HeaderController.module.scss
@@ -1,0 +1,8 @@
+.header {
+ &_controller {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  grid-column-gap: 50px;
+ }
+}
