+++ src/components/Home/Home.module.scss
@@ -1,0 +1,3 @@
+.home {
+ background: url('../../assets/poster_1.png');
+}
