+++ src/components/index.tsx
@@ -2,10 +2,6 @@
 export * from './Container/Container';
 export * from './Header/Header';
 export * from './Home/Home';
-export * from './Header/HeaderHome/HeaderHome';
-export * from './Header/HeaderForm/HeaderForm';
-export * from './Header/HeaderControl/HeaderControl';
-export * from './Header/HeaderController/HeaderController';
 
 //UI
 export * from '../UI/Input/Input';
