--- src/styles/_fonts.scss
@@ -1,1 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=DM+Sans:wght@400;500;700&display=swap');
