--- src/styles/parts/_header.scss
@@ -1,46 +1,0 @@
-.header {
-    padding: 15px 0;
-    &-wrapper {
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-    }
-    &-link {
-        width: 185px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        grid-column-gap: 24px;
-        cursor: pointer;
-        text-decoration: none;
-        &:hover {
-            .header-name {
-                color: #be123c;
-            }
-        }
-    }
-    &-name {
-        font-weight: 700;
-        font-size: 24px;
-        line-height: 24px;
-        color: #ffffff;
-        transition: 0.3s;
-    }
-    &-control {
-        font-weight: 700;
-        font-size: 16px;
-        line-height: 24px;
-        color: #ffffff;
-        cursor: pointer;
-        transition: 0.3s;
-        &:hover {
-            color: #be123c;
-        }
-    }
-    &-controller {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        grid-column-gap: 50px;
-    }
-}
