--- src/styles/parts/_home.scss
@@ -1,3 +1,0 @@
-.home {
- background: url('../assets/poster_1.png');
-}
