+++ src/styles/style.scss
@@ -1,6 +1,4 @@
-@import 'normalize';
-@import 'fonts';
-@import 'settings';
-
-@import './parts/header';
-@import './parts/home';
+@import url('https://fonts.googleapis.com/css2?family=DM+Sans:wght@400;500;700&display=swap');
+body {
+ font-family: 'DM Sans', sans-serif;
+}
